Temporary allow consensus index dance in CommandChunk as well ()

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/chunk/ChunkedTransaction.java
@@ -151,7 +151,7 @@ public class ChunkedTransaction implements CommandBatchToApply {
                 beforeTransactionStartPosition,
                 positionAfter,
                 checksum,
-                chunk.chunkMetadata().consensusIndex());
+                chunk.chunkMetadata().consensusIndex().longValue());
         lastBatchLogPosition = beforeStart;
     }
